Mirror the recent Alex's change for Windows. It doesn't seem being strictly required, but let's be safe.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/branches/B2_5_Release@51332 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/isc_sync.cpp
@@ -2397,7 +2397,7 @@ void ISC_unmap_object(ISC_STATUS* status_vector,
  **************************************/
 	SYSTEM_INFO sys_info;
 	GetSystemInfo(&sys_info);
-	const ULONG page_size = sys_info.dwAllocationGranularity;
+	const size_t page_size = sys_info.dwAllocationGranularity;
 
 	// Compute the start and end page-aligned offsets which
 	// contain the object being mapped.
